requirements.txt
boto3==1.26.142
botocore==1.29.142
gTTS==2.3.2
moviepy==1.0.3
playwright==1.34.0
praw==7.7.0
prawcore~=2.3.0
requests==2.31.0
rich==13.4.1
toml==0.10.2
translators==5.7.6
pyttsx3==2.90
Pillow==9.5.0
tomlkit==0.11.8
Flask==2.3.3
clean-text==0.6.0
unidecode==1.3.6
spacy==3.5.3
torch==2.0.1
transformers==4.29.2
ffmpeg-python==0.2.0
elevenlabs==0.2.17
yt-dlp==2023.7.6
better_profanity==0.7.0